+++ com.unity.visualeffectgraph/CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [4.10.0-preview] - 2019-XX-XX
+- Fixed error spam when opening a vfx with the visual effect tab hidden
 
 ## [4.9.0-preview] - 2019-01-28